backend/pyproject.toml
[project]
name = "threatweaver-backend"
version = "0.1.0"
description = "ThreatWeaver Multi-Agent Cybersecurity Platform - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "greenlet>=3.1.1",
    "alembic>=1.13.3",
    "asyncpg>=0.29.0",
    "redis>=5.2.0",
    "slowapi>=0.1.9",
    "httpx>=0.27.2",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "litellm>=1.51.0",
    "langgraph>=0.2.45",
    "langchain>=0.3.7",
    "langchain-core>=0.3.15",
    "langchain-openai>=0.2.8",
    "email-validator>=2.3.0",
    "bcrypt>=5.0.0",
    "argon2-cffi>=25.1.0",
    "nexus-ai-fs>=0.5.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.2",
    "ruff>=0.7.4",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing"

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
]
